+++ test/unit/ConfigFactoryInterfaceTest.php
@@ -3,7 +3,7 @@
 namespace Dhii\Config\UnitTest;
 
 use Xpmock\TestCase;
-use Dhii\Config\ConfigAwareInterface as TestSubject;
+use Dhii\Config\ConfigFactoryInterface as TestSubject;
 use PHPUnit_Framework_MockObject_MockObject as MockObject;
 
 /**
@@ -11,26 +11,26 @@ use PHPUnit_Framework_MockObject_MockObject as MockObject;
  *
  * @since [*next-version*]
  */
-class ConfigAwareInterfaceTest extends TestCase
+class ConfigFactoryInterfaceTest extends TestCase
 {
     /**
      * The name of the test subject.
      *
      * @since [*next-version*]
      */
-    const TEST_SUBJECT_CLASSNAME = 'Dhii\Config\ConfigAwareInterface';
+    const TEST_SUBJECT_CLASSNAME = 'Dhii\Config\ConfigFactoryInterface';
 
     /**
      * Creates a new instance of the test subject.
      *
      * @since [*next-version*]
      *
-     * @return TestSubject|MockObject The new instance.
+     * @return TestSubject|MockObject The new
      */
     public function createInstance()
     {
         $mock = $this->getMockBuilder(static::TEST_SUBJECT_CLASSNAME)
-            ->setMethods(array('getConfig'))
+            ->setMethods(array('make'))
             ->getMock();
 
         return $mock;
@@ -46,5 +46,6 @@ class ConfigAwareInterfaceTest extends TestCase
         $subject = $this->createInstance();
 
         $this->assertInstanceOf(static::TEST_SUBJECT_CLASSNAME, $subject, 'A valid instance of the test subject could not be created');
+        $this->assertInstanceOf('Dhii\Factory\FactoryInterface', $subject, 'Subject does not implement required interface');
     }
 }